+++ max_div/_cli/__init__.py
@@ -2,5 +2,7 @@ from ._cli import *
 from ._cmd_benchmark import *
 from ._cmd_benchmark_internal import *
 from ._cmd_benchmark_solver import *
+from ._cmd_benchmark_solver_list_problems import *
+from ._cmd_benchmark_solver_strategies import *
 from ._cmd_numba_status import *
 from ._cmd_solve import *