+++ max_div/_cli/bm_solver_strategies/_base.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections import defaultdict
-from functools import cached_property
 
 from tqdm import tqdm